address delete page

+++ app/Http/Controllers/UserPaymentConfigController.php
@@ -45,9 +45,17 @@ class UserPaymentConfigController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(UserPaymentConfig $userPaymentConfig, Request $request)
     {
-        //
+        $user = auth()->user();
+        if($userPaymentConfig['user_id'] === $user->id)
+        {
+            //
+        }
+        else
+        {
+            return abort(404);
+        }
     }
 
     /**
@@ -69,10 +77,18 @@ class UserPaymentConfigController extends Controller
      */
     public function edit(UserPaymentConfig $userPaymentConfig)
     {
-        return view('user_payment_config.edit', [
+        $user = auth()->user();
+        if($userPaymentConfig['user_id'] === $user->id)
+        {
+            return view('user_payment_config.edit', [
             'title' => 'Edit Billing Card',
             'billingCard' => $userPaymentConfig,
         ]);
+        }
+        else
+        {
+            return abort(404);
+        }
     }
 
     /**
@@ -85,82 +101,96 @@ class UserPaymentConfigController extends Controller
     public function update(Request $request, UserPaymentConfig $userPaymentConfig)
     {
         $user = auth()->user();
-
-        $validator = Validator::make($request->all(), [
-            'card_holder_name' => 'required|min: 6|max: 191',
-            'card_number' => 'required|digits: 16',
-            'expiry_date' => 'required', // format 2018-01
-
-            'building_name' => 'required|string|max:191',
-            'street_address1' => 'required|max:191',
-            'street_address2' => 'max:191',
-            'street_address3' => 'max:191',
-            'street_address4' => 'max:191',
-            'postcode' => 'required|string|min: 5|max:191',
-            'city' => 'required|string|min: 4|max:191',
-            'country' => 'required|string|min: 4|max:191',
-            'phone_number_extension' => 'required|min: 2|max:191',
-            'phone_number' => 'required|min: 5|max:191',
-            'mobile_number_extension' => 'max:191',
-            'mobile_number' => 'max:191',
-        ]);
-
-        if(empty($validator->errors()->all()))
+        if($userPaymentConfig['user_id'] === $user->id)
         {
-            $expiry_date = explode('-', $request->input('expiry_date'));
-            $expiry_year = $expiry_date[0];
-            $expiry_month = $expiry_date[1];
+            $validator = Validator::make($request->all(), [
+                'card_holder_name' => 'required|min: 6|max: 191',
+                'card_number' => 'required|digits: 16',
+                'expiry_date' => 'required', // format 2018-01
 
-            if(strtotime(date("$expiry_year-$expiry_month")) >= strtotime(date('Y-m')))
+                'building_name' => 'required|string|max:191',
+                'street_address1' => 'required|max:191',
+                'street_address2' => 'max:191',
+                'street_address3' => 'max:191',
+                'street_address4' => 'max:191',
+                'postcode' => 'required|string|min: 5|max:191',
+                'city' => 'required|string|min: 4|max:191',
+                'country' => 'required|string|min: 4|max:191',
+                'phone_number_extension' => 'required|min: 2|max:191',
+                'phone_number' => 'required|min: 5|max:191',
+                'mobile_number_extension' => 'max:191',
+                'mobile_number' => 'max:191',
+            ]);
+
+            if(empty($validator->errors()->all()))
             {
-                $data = array(
-                    'card_holder_name' => request('card_holder_name'),
-                    'card_number' => request('card_number'),
-                    'expiry_month' => $expiry_month,
-                    'expiry_year' => $expiry_year,
+                $expiry_date = explode('-', $request->input('expiry_date'));
+                $expiry_year = $expiry_date[0];
+                $expiry_month = $expiry_date[1];
 
-                    'building_name' => request('building_name'),
-                    'street_address1' => request('street_address1'),
-                    'street_address2' => request('street_address2'),
-                    'street_address3' => request('street_address3'),
-                    'street_address4' => request('street_address4'),
-                    'postcode' => request('postcode'),
-                    'city' => request('city'),
-                    'country' => request('country'),
-                    'county' => request('county'), // nullable
-                    'phone_number_extension' => request('phone_number_extension'),
-                    'phone_number' => request('phone_number'),
-                    'mobile_number_extension' => request('mobile_number_extension'), // nullable
-                    'mobile_number' => request('mobile_number'), // nullable
-                );
+                if(strtotime(date("$expiry_year-$expiry_month")) >= strtotime(date('Y-m')))
+                {
+                    $data = array(
+                        'card_holder_name' => request('card_holder_name'),
+                        'card_number' => request('card_number'),
+                        'expiry_month' => $expiry_month,
+                        'expiry_year' => $expiry_year,
 
-                UserPaymentConfig::where('id', $userPaymentConfig->id)->update($data);
+                        'building_name' => request('building_name'),
+                        'street_address1' => request('street_address1'),
+                        'street_address2' => request('street_address2'),
+                        'street_address3' => request('street_address3'),
+                        'street_address4' => request('street_address4'),
+                        'postcode' => request('postcode'),
+                        'city' => request('city'),
+                        'country' => request('country'),
+                        'county' => request('county'), // nullable
+                        'phone_number_extension' => request('phone_number_extension'),
+                        'phone_number' => request('phone_number'),
+                        'mobile_number_extension' => request('mobile_number_extension'), // nullable
+                        'mobile_number' => request('mobile_number'), // nullable
+                    );
 
-                return redirect()->route('billingHome')->with('flashSuccess', 'Billing card successfully updated.');
+                    UserPaymentConfig::where('id', $userPaymentConfig->id)->update($data);
+
+                    return redirect()->route('billingHome')->with('flashSuccess', 'Billing card successfully updated.');
+                }
+                else
+                {
+                    return redirect()->back()->with([
+                        'errors' => ['Invalid expiry date provided.'],
+                        'input' => $request->all(),
+                    ]);
+                }
             }
             else
             {
                 return redirect()->back()->with([
-                    'errors' => ['Invalid expiry date provided.'],
+                    'errors' => $validator->errors()->all(),
                     'input' => $request->all(),
                 ]);
             }
         }
         else
         {
-            return redirect()->back()->with([
-                'errors' => $validator->errors()->all(),
-                'input' => $request->all(),
-            ]);
+            return abort(404);
         }
     }
 
     public function delete(UserPaymentConfig $userPaymentConfig)
     {
-        return view('user_payment_config.delete', [
-            'title' => 'Delete Billing Card',
-            'billingCard' => $userPaymentConfig,
-        ]);
+        $user = auth()->user();
+        if($userPaymentConfig['user_id'] === $user->id)
+        {
+            return view('user_payment_config.delete', [
+                'title' => 'Delete Billing Card',
+                'billingCard' => $userPaymentConfig,
+            ]);
+        }
+        else
+        {
+            return abort(404);
+        }
     }
 
     /**
@@ -171,30 +201,38 @@ class UserPaymentConfigController extends Controller
      */
     public function destroy(UserPaymentConfig $userPaymentConfig, Request $request)
     {
-        $validator = Validator::make($request->all(), [
-            'choice' => 'required|boolean',
-        ]);
-
         $user = auth()->user();
-
-        if(empty($validator->errors()->all()))
+        if($userPaymentConfig['user_id'] === $user->id)
         {
-            $choice = (bool) $request->input('choice');
+            $validator = Validator::make($request->all(), [
+                'choice' => 'required|boolean',
+            ]);
 
-            if($choice !== FALSE)
+            $user = auth()->user();
+
+            if(empty($validator->errors()->all()))
             {
-                UserPaymentConfig::destroy($userPaymentConfig->id);
+                $choice = (bool) $request->input('choice');
 
-                return redirect()->route('billingHome')->with('flashSuccess', 'Billing card has been deleted successfully.');
+                if($choice !== FALSE)
+                {
+                    UserPaymentConfig::destroy($userPaymentConfig->id);
+
+                    return redirect()->route('billingHome')->with('flashSuccess', 'Billing card has been deleted successfully.');
+                }
+                else
+                {
+                    return redirect()->route('billingHome')->with('flashSuccess', 'Billing card has not been deleted.');
+                }
             }
             else
             {
-                return redirect()->route('billingHome')->with('flashSuccess', 'Billing card has not been deleted.');
+                return redirect()->back()->with('errors', $validator->errors()->all());
             }
         }
         else
         {
-            return redirect()->back()->with('errors', $validator->errors()->all());
+            return abort(404);
         }
     }
 }
